requirements.txt
selene>=2.0.0rc7
pytest>=7.4.3
allure-pytest==2.13.3
python-dotenv==1.0
requests>=2.31.0
pypandoc_binary==1.13
jsonschema